LayoutTests/fast/dom/node-prefix-setter-namespace-exception-expected.txt

Test how Node.prefix setter raises NAMESPACE_ERR.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS href.prefix is null
PASS document.createAttribute('attr').prefix = 'abc' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createAttributeNS(null, 'attr').prefix = 'abc' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createElementNS(null, 'attr').prefix = 'abc' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createAttributeNS('foo', 'bar').prefix = 'xml' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createElementNS('foo', 'bar').prefix = 'xml' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createAttribute('attr').prefix = 'xmlns' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createAttributeNS('foo', 'attr').prefix = 'xmlns' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS document.createAttribute('xmlns').prefix = 'foo' threw exception NamespaceError (DOM Exception 14): The operation is not allowed by Namespaces in XML..
PASS successfullyParsed is true

TEST COMPLETE
